Code for list of prime no. from 1 to n.
This source code includes the methods or functions to calculate the list of all prime numbers from 1 to n. Each function is an improved form of the above function.
The sieve method is the most optimized function to find the list of prime numbers in this source code yet. A better-optimized function in terms of space complexity will be published soon.